refactor(frontend): refactor project structure
