+++ frontend/src/components/SocialList.tsx
@@ -5,6 +5,7 @@ import { userApi } from "@/api/user"
 import { User } from "@/types/api"
 import UserAvatar from "./UserAvatar"
 import { useAuthState } from "@/context/auth"
+import { A } from "@solidjs/router"
 
 export default function SocialList() {
     const authState = useAuthState()
@@ -66,24 +67,26 @@ export default function SocialList() {
         <div class="space-y-3">
             <For each={follows() || []}>
                 {(user) => (
-                    <div class="flex items-start">
-                        <div class="relative">
-                            {/* Avatar */}
-                            <UserAvatar user={user} w={10} h={10} />
-                            <div
-                                class={`absolute -bottom-0.5 -right-0.5 w-3 h-3 rounded-full ${user.presence?.status === "online" ? "bg-base0B" : "bg-base04"}`}
-                            />
-                        </div>
+                    <A href={`/users/${user.id}`}>
+                        <div class="flex items-start">
+                            <div class="relative">
+                                {/* Avatar */}
+                                <UserAvatar user={user} w={10} h={10} />
+                                <div
+                                    class={`absolute -bottom-0.5 -right-0.5 w-3 h-3 rounded-full ${user.presence?.status === "online" ? "bg-base0B" : "bg-base04"}`}
+                                />
+                            </div>
 
-                        {/* Presence */}
-                        <div class="flex-1 ml-2 min-w-0">
-                            <p class="text-sm">{user.username}</p>
-                            <p class="text-sm truncate">
-                                {formatActivity(user.presence?.activityType, user.presence?.activityName)}
-                            </p>
-                            <p class="text-xs text-base04 mt-1">{timeAgo(user.presence?.activityTimestamp)}</p>
+                            {/* Presence */}
+                            <div class="flex-1 ml-2 min-w-0">
+                                <p class="text-sm">{user.username}</p>
+                                <p class="text-sm truncate">
+                                    {formatActivity(user.presence?.activityType, user.presence?.activityName)}
+                                </p>
+                                <p class="text-xs text-base04 mt-1">{timeAgo(user.presence?.activityTimestamp)}</p>
+                            </div>
                         </div>
-                    </div>
+                    </A>
                 )}
             </For>
         </div>
