+++ frontend/src/components/UserAvatar.tsx
@@ -6,6 +6,7 @@ type UserAvatarProps = {
     w: number
     h: number
     onAvatarChange?: (f: File) => void
+    isCurrentUser?: boolean
 }
 
 export default function UserAvatar(props: UserAvatarProps) {
@@ -22,7 +23,7 @@ export default function UserAvatar(props: UserAvatarProps) {
         }
     `
     return (
-        <div class="relative">
+        <div class="relative m-auto">
             <Show
                 when={props.user.avatarUrl}
                 fallback={
@@ -35,8 +36,8 @@ export default function UserAvatar(props: UserAvatarProps) {
                     <img class="w-full h-full object-cover" src={props.user.avatarUrl} alt="User avatar" />
                 </div>
             </Show>
-            <Show when={props.onAvatarChange}>
-                <label class="absolute rounded bottom-3 right-0 text-xs px-2 py-1 cursor-pointer bg-base02 hover:bg-base03">
+            <Show when={props.onAvatarChange && props.isCurrentUser}>
+                <label class="absolute rounded bottom-0 right-0 text-xs px-2 py-1 cursor-pointer bg-base02 hover:bg-base03">
                     Change
                     <input
                         type="file"
